requirements.txt
pykeepass
notify-send
git+https://github.com/bcbnz/python-rofi.git
